Minor cleanup on these classes
